Can restore.

+++ settings.py
@@ -1,6 +1,6 @@
 import json
 import os
-from os.path import join
+from os.path import join, isfile
 
 from subprocess import check_output
 
@@ -51,15 +51,16 @@ def get_secret(name):
 
 
 def save_secrets(get_passphrase):
-    secrets = {
-        "aws_access_key_id": get_secret("aws_access_key_id"),
-        "aws_secret_access_key": get_secret("aws_secret_access_key"),
-    }
-    if get_passphrase:
-        secrets["passphrase"] = get_secret("passphrase")
+    if not isfile(secrets_path):
+        secrets = {
+            "aws_access_key_id": get_secret("aws_access_key_id"),
+            "aws_secret_access_key": get_secret("aws_secret_access_key"),
+        }
+        if get_passphrase:
+            secrets["passphrase"] = get_secret("passphrase")
 
-    with open(secrets_path, 'a'):  # Create file if does not exist
-        pass
-    os.chmod(secrets_path, 0o600)
-    with open(secrets_path, 'w') as f:
-        json.dump(secrets, f)
+        with open(secrets_path, 'a'):  # Create file if does not exist
+            pass
+        os.chmod(secrets_path, 0o600)
+        with open(secrets_path, 'w') as f:
+            json.dump(secrets, f)
